requirements.txt
pillow==11.2.1
pydicom==3.0.1
urwid==3.0.2
wcwidth==0.2.13
